// src/types/text.ts
import { Document } from "@contentful/rich-text-types";

export interface IPulseText {
  fields: {
    childtitle: Document;
    title: string;
    slug: string;
    content: Document;
    conTitle: string;
    content1: string;
    contitle2: string;
    content2: string;
    contitle3: string;
    content3: string;
    contitle4: string;
    content4: string;
  };
}